requirements.txt
alembic==1.13.3
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
decorator==5.1.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.0.1
Faker==14.2.0
Flask>=2.2.5,<3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy>=3.1.1

greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
migrate==0.3.8
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==4.2.2
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.2.7
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.8.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.23
SQLAlchemy-serializer==1.4.1
stack-data==0.6.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.2.3
wcwidth==0.2.9
Werkzeug==2.2.2
WTForms==3.1.2
zipp==3.20.0
